project_elite_data_eng/requirements.txt
awswrangler==3.11.0
gspread == 6.2.1
apache-airflow-providers-amazon
apache-airflow-providers-postgres
faker==37.4.2
